+++ test/ziggurat/kafka_delay_test.clj
@@ -1,18 +1,27 @@
 (ns ziggurat.kafka-delay-test
   (:require [clojure.test :refer :all]
             [ziggurat.kafka-delay :refer :all]
-            [ziggurat.util.time :refer [get-current-time-in-millis]]
-            [ziggurat.metrics :as metrics]))
+            [ziggurat.metrics :as metrics]
+            [ziggurat.util.time :refer [get-current-time-in-millis]]))
 
 (deftest calculate-and-report-kafka-delay-test
-  (testing "calculates and reports the timestamp delay"
-    (let [record-timestamp 1528720767777
-          current-time     1528720768777
-          expected-delay   1000
-          namespace        "test"]
-      (with-redefs [get-current-time-in-millis (constantly current-time)
-                    metrics/report-time        (fn [metric-namespace delay]
-                                                 (is (= delay expected-delay))
-                                                 (is (= metric-namespace namespace)))]
-        (calculate-and-report-kafka-delay namespace record-timestamp)))))
-
+  (let [record-timestamp    1528720767777
+        current-time        1528720768777
+        expected-delay      1000
+        expected-namespaces ["test"]]
+    (testing "calculates and reports the timestamp delay"
+      (let [expected-additional-tags {:topic_name "expected-topic-entity-name"}]
+        (with-redefs [get-current-time-in-millis (constantly current-time)
+                      metrics/report-time        (fn [metric-namespaces delay additional-tags]
+                                                   (is (= delay expected-delay))
+                                                   (is (= metric-namespaces expected-namespaces))
+                                                   (is (= additional-tags expected-additional-tags)))]
+          (calculate-and-report-kafka-delay expected-namespaces record-timestamp expected-additional-tags))))
+    (testing "calculates and reports the timestamp delay when additional tags is empty or nil"
+      (let [expected-additional-tags nil]
+        (with-redefs [get-current-time-in-millis (constantly current-time)
+                      metrics/report-time        (fn [metric-namespaces delay additional-tags]
+                                                   (is (= delay expected-delay))
+                                                   (is (= metric-namespaces expected-namespaces))
+                                                   (is (= additional-tags expected-additional-tags)))]
+          (calculate-and-report-kafka-delay expected-namespaces record-timestamp))))))