+++ CHANGELOG.md
@@ -3,8 +3,20 @@ All notable changes to this project will be documented in this file. This change
 
 ## Unreleased Changes
 
-## 3.0.0-rc1 - 2019-05-17
+
+## 3.0.0-alpha - 2019-06-21
 - Upgrades kafka streams to version 2.1. Please refer [this](UpgradeGuide.md) to upgrade
+
+## 2.12.0 - 2019-06-17
+- Add support for providing a topic-name label in the metrics
+- Multiple Kafka producers support in ziggurat (#55)
+- Validate stream routes only when modes is not present or it contains stream-server (#59)
+
+## 2.11.1 - 2019-06-04
+- Actor stop fn should stop before the Ziggurat state (#53)
+
+## 2.11.0 - 2019-05-31
+- Running ziggurat in different modes (#46)
 
 ## 2.10.2 - 2019-05-03
 - Adds config to change the changelog topic replication factor